Break tests into smaller modules
