+++ tests/src/test_projects/merkle_proof/tests/functions/leaf_digest.rs
@@ -1,0 +1,29 @@
+use crate::merkle_proof::tests::utils::{
+    abi_calls::leaf_digest,
+    test_helpers::{build_tree, merkle_proof_instance},
+};
+
+mod leaf_digest {
+
+    use super::*;
+
+    mod succes {
+
+        use super::*;
+
+        #[ignore]
+        #[tokio::test]
+        async fn computes_leaf() {
+            let instance = merkle_proof_instance().await;
+
+            let leaves = ["A".as_bytes(), "B".as_bytes(), "C".as_bytes()].to_vec();
+            let key = 0;
+            let (_tree, _root, leaf, _proof) = build_tree(leaves, key).await;
+
+            assert_eq!(
+                leaf_digest(&instance, "A".as_bytes().try_into().unwrap()).await,
+                leaf
+            );
+        }
+    }
+}
