+++ tests/src/test_projects/merkle_proof/tests/functions/mod.rs
@@ -1,0 +1,4 @@
+mod leaf_digest;
+mod node_digest;
+mod process_proof;
+mod verify_proof;
