+++ tests/src/test_projects/merkle_proof/tests/mod.rs
@@ -1,0 +1,2 @@
+mod functions;
+mod utils;
